Introduce control.monad
...which adds the Maybe type and Functor, Applicative and Monad
instances for Maybe and Array. Also adds Show instances for those
two types and therefore moves and renames the showList function
from prelude.js.
"Typeclass"-functions are implemented as methods on the prototypes of
the respective types; however wrapper functions allow these to be called
in a functional way* (preserving polymorphism).

*e.g. fmap (function) (functor) instead of functor.fmap (function)