Improve implementation of Maybe
Make use of the new module meta.util (see last commit).
Most type class methods now need to be implemented for each value
constructor separately (Just, Nothing), not for the type (Maybe).
Example:
    // instance Functor Maybe
    Just.prototype.fmap = function (f) { ...
    Nothing.prototype.fmap = function (f) { ...
Some methods will still be added to the type directly (i.e. not on the
prototype).
Example:
    // instance Monad Maybe
    Maybe.pure = Just;
The .valueconstructor property on the prototype of new values allows to
check what value constructor a value was created with.
Example:
    const myMaybe = Just (1);
    myMaybe.valueconstructor.name === "Just" // true
    Nothing.valueconstructor.name === "Nothing" // false
The newly introduced functions isJust and isNothing make use of this.
Example:
    isJust (myMaybe); // true
    isNothing (myMaybe); // false